Implemented the new JSON format of TNHTweaker 1.2.0
Still keeping the old format because i want to be able to convert old characters to the new format.
