+++ src/TNHTweaker-UI.BusinessLogic/Interfaces/ICharacterSerializer.cs
@@ -8,18 +8,18 @@ namespace TNHTweaker_UI.BusinessLogic.Interfaces
     public interface ICharacterSerializer
     {
         /// <summary>
-        /// Attempts to read in a custom character from a text string, typically the contents of a custom character text file.
-        /// Reads, parses and puts the data into a <see cref="CharacterInfo"/> object.
+        /// Attempts to read in a custom character from a text string, typically the contents of a custom character text file using the new JSON format.
+        /// Reads, parses and puts the data into a <see cref="CharacterInformation"/> object.
         /// </summary>
-        /// <param name="characterDefinition">The full character definition to parse.</param>
-        /// <returns>A <see cref="CharacterInfo"/> instance with the given <paramref name="characterDefinition"/>. Or <c>null</c> if parsing encountered an error.</returns>
-        CharacterInfo ReadCharacterFromString(string[] characterDefinition);
+        /// <param name="characterDefinition">The full character definition to parse in JSON format.</param>
+        /// <returns>A <see cref="CharacterInformation"/> instance with the given <paramref name="characterDefinition"/>. Or <c>null</c> if parsing encountered an error.</returns>
+        CharacterInformation ReadCharacterFromString(string characterDefinition);
 
         /// <summary>
-        /// Writes the given <paramref name="character"/> as a text string that can be written to a customer character text file.
+        /// Writes the given <paramref name="character"/> as a text string that can be written to a customer character text file using the new JSON format.
         /// </summary>
-        /// <param name="character">The <see cref="CharacterInfo"/> instance to write out as a string definition.</param>
+        /// <param name="character">The <see cref="CharacterInformation"/> instance to write out as a string definition.</param>
         /// <returns>A custom character string notation of the given <paramref name="character"/>.</returns>
-        string WriteCharacterToString(CharacterInfo character);
+        string WriteCharacterToString(CharacterInformation character);
     }
 }
