+++ src/TNHTweaker-UI.BusinessLogic/TNHTweaker-UI.BusinessLogic.csproj
@@ -6,10 +6,17 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\TNHTweaker-UI.Models\TNHTweaker-UI.Models.csproj" />
   </ItemGroup>
 
   <ItemGroup>
+    <None Update="character.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
     <None Update="character.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
