+++ src/TNHTweaker-UI.Models/OldFormat/LevelEntry.cs
@@ -1,7 +1,6 @@
-﻿using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using TNHTweaker_UI.Models.Attributes;
 
-namespace TNHTweaker_UI.Models
+namespace TNHTweaker_UI.Models.OldFormat
 {
     /// <summary>
     /// Class that defines the challenges for a single level in the progression for a character.
@@ -11,39 +10,29 @@ namespace TNHTweaker_UI.Models
         /// <summary>
         /// The number of tokens the character is rewarded for completing the level's <see cref="TakeChallenge"/>.
         /// </summary>
-        [JsonProperty("NumOverrideTokensForHold")]
+        [PropertyName("NumOverrideTokensForHold")]
         public int NumberOfOverrideTokensForHold { get; set; }
 
         /// <summary>
         /// The number of additional supply points that will be spawned on top of the normal amount.
         /// </summary>
+        [PropertyName("@AdditionalSupplyPoints")]
         public int AdditionalSupplyPoints { get; set; }
 
-        /// <summary>
-        /// The maximum amount of supply boxes to spawn.
-        /// </summary>
+        [PropertyName("@MaxBoxesSpawned")]
         public int MaxBoxesSpawned { get; set; }
 
-        /// <summary>
-        /// The minimum amount of supply boxes to spawn.
-        /// </summary>
+        [PropertyName("@MinBoxesSpawned")]
         public int MinBoxesSpawned { get; set; }
 
-        /// <summary>
-        /// The maximum amount of override tokens that can be found in supply boxes per supply point.
-        /// </summary>
+        [PropertyName("@MaxTokensPerSupply")]
         public int MaxTokensPerSupply { get; set; }
 
-        /// <summary>
-        /// The minimum amount of override tokens that can be found in supply boxes per supply point.
-        /// </summary>
+        [PropertyName("@MinTokensPerSupply")]
         public int MinTokensPerSupply { get; set; }
 
-        /// <summary>
-        /// The percentage chance that a supply token can spawn when breaking a supply box.
-        /// Accepts a value between 0 and 1.
-        /// </summary>
-        public double BoxTokenChance { get; set; }
+        [PropertyName("@BoxTokenChance")]
+        public float BoxTokenChance { get; set; }
 
         /// <summary>
         /// Definition for defenses that are spawned at the hold point.
@@ -51,9 +40,9 @@ namespace TNHTweaker_UI.Models
         public TakeSupplyChallenge TakeChallenge { get; set; }
 
         /// <summary>
-        /// Definition for all phases of this hold.
+        /// Definition for encryptions and attacking waves during a hold.
         /// </summary>
-        public IList<PhaseDefinition> HoldPhases { get; set; }
+        public HoldChallenge HoldChallenge { get; set; }
 
         /// <summary>
         /// Definition for defenses that are spawned at supply points.
@@ -63,6 +52,11 @@ namespace TNHTweaker_UI.Models
         /// <summary>
         /// Definition for the patrols that spawn before the hold.
         /// </summary>
-        public IList<PatrolDefinition> Patrols { get; set; }
+        public PatrolChallenge PatrolChallenge { get; set; }
+
+        /// <summary>
+        /// Definition for traps that spawn before the hold.
+        /// </summary>
+        public TrapsChallenge TrapsChallenge { get; set; }
     }
 }
