+++ src/TNHTweaker-UI.Models/OldFormat/PatrolChallenge.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TNHTweaker_UI.Models
+namespace TNHTweaker_UI.Models.OldFormat
 {
     /// <summary>
     /// Class that defines the patrols that will be roaming around between holds.
