+++ src/TNHTweaker-UI.Models/OldFormat/PhaseDefinition.cs
@@ -1,8 +1,11 @@
-﻿using Newtonsoft.Json;
+﻿using TNHTweaker_UI.Models.Attributes;
 using TNHTweaker_UI.Models.Enums;
 
-namespace TNHTweaker_UI.Models
+namespace TNHTweaker_UI.Models.OldFormat
 {
+    /// <summary>
+    /// Class that defines the options for a single phase of a hold.
+    /// </summary>
     public class PhaseDefinition
     {
         /// <summary>
@@ -10,15 +13,9 @@ namespace TNHTweaker_UI.Models
         /// </summary>
         public EncryptionType Encryption { get; set; }
 
-        /// <summary>
-        /// The type of sosig to spawn as normal enemies in this phase.
-        /// </summary>
-        public string EnemyType { get; set; }
-
-        /// <summary>
-        /// The type of sosig to spawn as a leader enemy in this phase.
-        /// </summary>
-        public string LeaderType { get; set; }
+        //TODO Not sure what these types do yet. seems to define two different enemy types that can spawn. But i don't know the difference between EType and LType.
+        public string EType { get; set; }
+        public string LType { get; set; }
 
         /// <summary>
         /// The minimum amount of encryption targets that will spawn at the end of the phase.
@@ -40,10 +37,8 @@ namespace TNHTweaker_UI.Models
         /// </summary>
         public int MaxEnemies { get; set; }
 
-        /// <summary>
-        /// The amount of time in between waves of enemies during a phase.
-        /// </summary>
-        public double SpawnCadence { get; set; }
+        //TODO Not quite sure what spawn cadence could mean. Maybe it means it's the number of seconds before the next wave is spawned? needs testing!
+        public int SpawnCadence { get; set; }
 
         /// <summary>
         /// The maximum amount of enemies that can be alive at once during the phase.
@@ -64,28 +59,25 @@ namespace TNHTweaker_UI.Models
         /// <summary>
         /// Probably the amount of time in seconds before enemies start spawning at the start of a phase.
         /// </summary>
-        public int WarmUpTime { get; set; }
+        public int WarmUp { get; set; }
 
         /// <summary>
         /// Sets the team that spawned attackers will be on. Default is 1, and the players IFF is 0.
         /// </summary>
-        [JsonProperty("IFFUsed")]
+        [PropertyName("IFFUsed")]
         public int IffUsed { get; set; } = 1;
 
         /// <summary>
         /// The chance that the enemies will throw a grenade at the player.
         /// Value between 0 and 1, which represents a percentage.
         /// </summary>
+        [PropertyName("@GrenadeChance")]
         public float GrenadeChance { get; set; }
 
         /// <summary>
         /// The type of grenade the enemies will throw at the player.
         /// </summary>
+        [PropertyName("@GrenadeType")]
         public string GrenadeType { get; set; }
-
-        /// <summary>
-        /// Determines if enemies should swarm the player.
-        /// </summary>
-        public bool SwarmPlayer { get; set; }
     }
 }
