+++ src/TNHTweaker-UI.Models/OldFormat/TrapDefinition.cs
@@ -1,6 +1,6 @@
 ﻿using TNHTweaker_UI.Models.Enums;
 
-namespace TNHTweaker_UI.Models
+namespace TNHTweaker_UI.Models.OldFormat
 {
     /// <summary>
     /// Class that defines a single trap type.
