+++ src/TNHTweaker-UI.Models/OldFormat/WeaponDefinition.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
+using TNHTweaker_UI.Models.Attributes;
 
-namespace TNHTweaker_UI.Models
+namespace TNHTweaker_UI.Models.OldFormat
 {
     /// <summary>
     /// Class that defines a weapon load-out.
@@ -13,23 +14,20 @@ namespace TNHTweaker_UI.Models
         public string[] ListOverride { get; set; } = new string[0];
 
         /// <summary>
-        /// Override what ammo types can be spawned from the ammo spawner.
-        /// </summary>
-        public IList<string> AmmoOverride { get; set; } = null;
-
-        /// <summary>
         /// Sets the number of magazines or speed loaders the weapon starts with.
         /// </summary>
-        public int NumMags { get; set; }
+        [PropertyName("Num_Mags_SL_Clips")]
+        public int NumMagsSlClips { get; set; }
 
         /// <summary>
         /// Sets the number of bullets a weapon starts with (possibly unused if spawned weapon takes magazines).
         /// </summary>
+        [PropertyName("Num_Rounds")]
         public int NumRounds { get; set; }
 
         /// <summary>
         /// A list of object tables which are used to determine what equipment could possibly spawn.
         /// </summary>
-        public IList<ObjectTableDefinition> Tables { get; set; } = new List<ObjectTableDefinition>();
+        public IList<ObjectTableDefinition> TableDefs { get; set; } = new List<ObjectTableDefinition>();
     }
 }
