+++ src/TNHTweaker-UI.Models/TNHTweaker-UI.Models.csproj
@@ -5,4 +5,8 @@
     <RootNamespace>TNHTweaker_UI.Models</RootNamespace>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+  </ItemGroup>
+
 </Project>
